made mask long long

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -5,7 +5,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned long int i = n;
+	unsigned long long i = n;
 
 	while (i > 0)
 	{
